Sufficient knowledge of God outside of God's gracious turning towards man in Jesus Christ isn't possible.
Despite all the depravity therefore, God yet testifies Himself so to the
Gentiles that if they do not want to accept Him, they stands guilty
before God.

Not for one moment the knowledge of God in Romans 1, compares to or
constitutes a play of the pure and true knowledge of God that can only
be found through Jesus Christ in the Scriptures.